Refactorings in the roundup operator
