+++ src/compute/src/sink/mod.rs
@@ -8,4 +8,5 @@
 // by the Apache License, Version 2.0.
 
 mod persist_sink;
+mod refresh;
 mod subscribe;
